requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
crewai==0.28.8
groq==0.4.2
langchain-groq==0.0.1
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2 